refactor utilities classes semantic
