+++ backend/pgdas_fiscal_oesk/defis.py
@@ -8,6 +8,7 @@ from pgdas_fiscal_oesk import Consultar
 from pgdas_fiscal_oesk.defis_utils.legato import Legato
 # from pgdas_fiscal_oesk.defis_utils.legato import transformers as tfms
 from simples_nacional_utilities import SimplesNacionalUtilities
+from compt_utils import get_compt
 
 COMPT = get_compt(-1)
 CONS = Consultar(COMPT)
