+++ backend/utilities/default/__init__.py
@@ -1,10 +1,10 @@
 from .interact import ativa_janela, press_keys_b4, press_key_b4, all_keys, tk_msg, _contmatic_select_by_name
 from .sets import HandleFileOperations
-from .sets import get_all_valores
 from .sets.init_email import EmailExecutor
+from .sets.pathmanager import HasJson
 
 # dates
-from .sets import calc_date_compt_offset, get_compt, compt_to_date_obj
+from .sets import trata_money_excel
 # drivers
 from .webdriver_utilities import pgdas_driver, pgdas_driver_ua, ginfess_driver, default_qrcode_driver, jucesp_simple_driver, proffile_noqr_driver
 # selenium shortcuts
@@ -17,3 +17,8 @@ from selenium.webdriver.support.wait import WebDriverWait
 from selenium.common.exceptions import *
 from selenium.webdriver import Remote
 from selenium import webdriver
+from selenium.webdriver import ActionChains
+
+# others
+import os
+from time import sleep
