+++ backend/utilities/default/sets/__init__.py
@@ -4,137 +4,74 @@ from utilities.default.sets.initial_set_files_location import InitialSetFilesLoc
 from .pathmanager import Dirs
 from .now import Now
 import os
+import glob
 
 
 # @ staticmethod
 
-
-def compt_to_date_obj(compt) -> date:
-    return Now.str_to_date(compt, "%m-%Y")
-
-
-def calc_date_compt_offset(m_cont=-1, y_cont=0, past_only=True, sep='-') -> date:
-    """ - returns `date` object based on the two first arguments
-    Args:
-        m_cont (int, optional): referes to month before/after now. Defaults to -1.
-        y_cont (int, optional): refers to year before/after now. Defaults to 0.
-        past_only (bool, optional): only returns the past, because of competencia. 
-        Defaults to True.
-    Returns:
-        date: datetime.date making the counters with relativedelta
-    """
-    from datetime import timedelta, datetime
-    from dateutil.relativedelta import relativedelta
-    month = datetime.now().month
-    year = datetime.now().year
-
-    now_date = date(year, month, 1)
-
-    if past_only:
-        m_cont = m_cont * (-1) if m_cont > 0 else m_cont
-        y_cont = y_cont * (-1) if y_cont > 0 else y_cont
-        # force to be negative
-
-    now_date = now_date + relativedelta(months=m_cont)
-    now_date = now_date + relativedelta(years=y_cont)
-    return now_date
+def trata_money_excel(faturado):
+    # TODO: refaturar em _backend com foco em já definir os valores e pegar do bd se tem DAS pendentes ou não
+    if faturado is None:
+        faturado = 0
+    faturado = float(faturado)
+    faturado = str(faturado).lower().strip()
+    if faturado == 'das_pend':
+        return 'ATENÇÃO, HÁ BOLETO(S) DO PGDAS PENDENTE(S)'
+    if 'nan' in faturado or 'zerou' in faturado or float(faturado) == 0:
+        faturado = '0, 00. SEM MOVIMENTO NESTE MÊS.'
+        return faturado
+    faturado = f'{float(faturado):,.2f}'
+    faturado = faturado.replace('.', 'v')
+    faturado = faturado.replace(',', '.')
+    faturado = faturado.replace('v', ',')
+    return faturado
 
 
-def get_compt(m_cont=-1, y_cont=0, past_only=True, format_str='-') -> str:
-    """ - returns `str` based on the two first arguments with the format option,
-    calls compt_as_date first, 
-    Args:
-        m_cont (int, optional): referes to month before/after now. Defaults to -1.
-        y_cont (int, optional): refers to year before/after now. Defaults to 0.
-        past_only (bool, optional): only returns the past, because of competencia. 
-        Defaults to True.
-        format (str, optional): if len(format) > 2 it retuns what is set on the string.
-        Defaults to 2 characters + '%m-%Y'
-    Returns:
-        str: A string representing the month and year in the specified format.
-    """
-    now_date = calc_date_compt_offset(m_cont, y_cont, past_only)
-    if len(format_str) <= 2:
-        return now_date.strftime(f'%m{format_str}%Y')
-    else:
-        return now_date.strftime(format_str)
-
-
-def get_all_valores(sem_ret, com_ret, anexo, valor_tot) -> list:
-    def split_sep_vals(val, withme=';'):
-        val = str(val)
-        try:
-            return val.split(withme)
-            # print('valorA;valorB___valorC;valorD  valorTotal')
-
-        except AttributeError:
-            return list(val)
-
-    def greater_than(l1, l2):
-        return len(l1) > len(l2)
-
-    sem_ret = split_sep_vals(sem_ret)
-    com_ret = split_sep_vals(com_ret)
-    anexo = split_sep_vals(anexo)
-
-    # retorna False para não prosseguir, pois o número de anexos não bate com o número de valores entre ;
-    if greater_than(sem_ret, com_ret):
-        if greater_than(anexo, sem_ret) or greater_than(sem_ret, anexo):
-            return False
-    elif greater_than(com_ret, sem_ret):
-        if greater_than(anexo, com_ret) or greater_than(com_ret, anexo):
-            return False
-    else:
-        # tanto faz porque com_ret == sem_ret
-        if greater_than(anexo, sem_ret) or greater_than(sem_ret, anexo):
-            return False
-
-    all_valores = []
-    soma_total = 0
-    for c in range(len(anexo)):
-        try:
-            sr = sem_ret[c]
-        except IndexError:
-            sr = 0
-        try:
-            cr = com_ret[c]
-        except IndexError:
-            cr = 0
-        # Se o valor não foi escrito, é considerado 0
-
-        anx = anexo[c]
-        try:
-            soma_total += float(sr) + float(cr)
-        except ValueError:
-            return False
-        all_valores.append({'valor_n_retido': sr,
-                            'valor_retido': cr, 'anexo': anx})
-
-    if float(soma_total) == float(valor_tot):
-        return all_valores
-    else:
-        print(soma_total, type(soma_total), valor_tot, type(valor_tot))
-        print('Vou retornar None')
-
-
+# TODO:
+# class FileOperations(Dirs, Now)
 class HandleFileOperations(Dirs, Now):
     files_location = InitialSetFilesLocation()
 
+    @staticmethod
+    def move_file(where_from, destiny):
+        """[File/folder moved from a place[where_from] to another[destiny]]
+        Args:
+            where_from (str):
+            destiny (str):
+        """
+        from shutil import move
+        move(where_from, destiny)
+
+    def certifs_exist(self, main_path, startswith, at_least=2, endswith: bool = False):
+        # if endswith is True, it will search for endswith instead
+        arqs_search = self.files_get_anexos_v4(main_path, 'png')
+        arqs_search += self.files_get_anexos_v4(main_path, 'pdf')
+        # certificados gias são em PDF...
+        arqs_search = [
+            self.path_leaf(f, True) for f in arqs_search]
+        if endswith is False:
+            arqs_search = [f for f in arqs_search if f.startswith(startswith)]
+        else:
+            arqs_search = [f for f in arqs_search if f.endswith(startswith)]
+
+        if len(arqs_search) >= at_least:
+            return True
+        return False
+
     @classmethod
-    def files_pathit(cls, pasta_client, insyear=None, ano=None):
+    def files_pathit(cls, pasta_client, insyear, ano=None):
         from dateutil import relativedelta as du_rl
 
         """[summary]
 
         Args:
             pasta_client (str): client folder name
-            insyear (str): inside year (competencia or whatever). Defaults then call get_compt_only() as default
+            insyear (str): inside year (competencia or whatever).
             ano (str,[optional]): year folder. Defaults to None.
 
         Returns:
             [type]: [description]
         """
-        insyear = get_compt() if insyear is None else insyear
         compt = insyear
         if ano is None:
             # ano = ''.join([insyear[e+1:] for e in range(len(insyear)) if insyear[e] not in '0123456789'])
@@ -186,50 +123,38 @@ class HandleFileOperations(Dirs, Now):
         if excluir_png:
             os.remove(filepath_png)
 
+    def files_get_anexos_v4(self, path, file_type='pdf', upload=True):
+        """
+        :param path: get anexos from the following path
+        :param file_type: file annexed type
+        :param upload: False -> email it! True: upload it!
+        :return: pdf_files or whatever
+
+        # _files_path
+        """
+        from email.mime.application import MIMEApplication
+        pdf_files = list()
+        # Lucas Restaurante
+
+        dir_searched_now = path
+        list_checked_returned = [os.path.join(dir_searched_now, fname)
+                                 for fname in os.listdir(dir_searched_now) if fname.lower().endswith(file_type)]
+
+        for fname in list_checked_returned:
+            if not upload:
+                file_opened = MIMEApplication(open(fname, 'rb').read())
+                fname_title = self.path_leaf(fname)
+                file_opened.add_header(
+                    'Content-Disposition', 'attachment', filename=fname_title)
+                pdf_files.append(file_opened)
+            else:
+                pdf_files.append(f'{fname}')
+        return pdf_files
+
     @staticmethod
-    def ate_atual_compt(compt_atual, first_compt=None):
-        from datetime import date
-        from dateutil import relativedelta
-        if first_compt is None or first_compt == compt_atual:
-            yield compt_atual
-        else:
-            first_compt = first_compt.split('-')
-            if len(first_compt) == 1:
-                first_compt = first_compt.split('/')
-            first_compt = [int(val) for val in first_compt]
-            first_compt = date(first_compt[1], first_compt[0], 1)
-
-            # next_date = first_compt + relativedelta.relativedelta(months=1)
-
-            last_compt = compt_atual.split('-')
-            # compt = [int(c) for c in compt]
-            last_compt = [int(v) for v in last_compt]
-            last_compt = date(last_compt[1], last_compt[0], 1)
-
-            # list_compts = []
-            while first_compt <= last_compt:
-                compt = first_compt
-                first_compt = first_compt + \
-                              relativedelta.relativedelta(months=1)
-                compt_appended = f'{compt.month:02d}-{compt.year}'
-                # list_compts.append(compt_appended)
-                yield compt_appended
-
-    @staticmethod
-    def trata_money_excel(faturado):
-        # TODO: refaturar em _backend com foco em já definir os valores e pegar do bd se tem DAS pendentes ou não
-        if faturado is None:
-            faturado = 0
-        faturado = float(faturado)
-        faturado = str(faturado).lower().strip()
-        if faturado == 'das_pend':
-            return 'ATENÇÃO, HÁ BOLETO(S) DO PGDAS PENDENTE(S)'
-        if 'nan' in faturado or 'zerou' in faturado or float(faturado) == 0:
-            faturado = '0, 00. SEM MOVIMENTO NESTE MÊS.'
-            return faturado
-        faturado = f'{float(faturado):,.2f}'
-        faturado = faturado.replace('.', 'v')
-        faturado = faturado.replace(',', '.')
-        faturado = faturado.replace('v', ',')
-        return faturado
-    # yield list_compts
+    def sort_files_by_most_recent(folderpath):
+        return sorted([os.path.join(folderpath, f)
+                       for f in os.listdir(folderpath)],
+                      key=lambda x: os.path.getmtime(
+                          os.path.join(folderpath, x)),
+                      reverse=True)
